mv utils functions to . . . . utils
